pyproject.toml
[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vllm-bench"
version = "0.1.0"
description = "A/B benchmark for MLA attention + quant fusion"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # vllm build-system requirements (from pyproject.toml)
    "cmake>=3.26.1",
    "ninja",
    "packaging>=24.2",
    "setuptools>=77.0.3,<81.0.0",
    "setuptools-scm>=8.0",
    "wheel",
    "jinja2",
    # runtime deps needed before vllm install
    "numpy",
    "pip",
    # torch with CUDA
    "torch==2.10.0",
    # benchmark CLI deps
    "pyyaml",
]

[project.scripts]
vllm-bench = "vllm_bench.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
torch = { index = "pytorch-cu130" }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
